builds/sonarr_cleaner/files/pip-requirements.txt
requests
logzero
plexapi
PyYAML>=5.2
